requirements.txt
Django==2.1.4
pytest-django
pytest
-e git+https://github.com/justinmayer/django-autoslug.git#egg=django-autoslug
django-templated-email
django-test-without-migrations
django-bootstrap4
Pillow
django-extra-views